fix: move testId to inputProps for all input components
