+++ packages/odyssey-react-mui/src/labs/VirtualizedAutocomplete.tsx
@@ -264,6 +264,7 @@ const VirtualizedAutocomplete = <
               ...params.inputProps,
               "aria-errormessage": errorMessageElementId,
               "aria-labelledby": labelElementId,
+              "data-se": testId,
             }}
             aria-describedby={ariaDescribedBy}
             id={id}
@@ -273,7 +274,7 @@ const VirtualizedAutocomplete = <
         )}
       />
     ),
-    [errorMessage, errorMessageList, hint, isOptional, label, nameOverride]
+    [errorMessage, errorMessageList, hint, isOptional, label, nameOverride, testId]
   );
   const onChange = useCallback<
     NonNullable<
@@ -313,7 +314,6 @@ const VirtualizedAutocomplete = <
       {...inputValueProp}
       // AutoComplete is wrapped in a div within MUI which does not get the disabled attr. So this aria-disabled gets set in the div
       aria-disabled={isDisabled}
-      data-se={testId}
       disableCloseOnSelect={hasMultipleChoices}
       disabled={isDisabled}
       freeSolo={isCustomValueAllowed}
